rename team.interface to team
